pyproject.toml
[tool.poetry]
name = "logger"
version = "0.1.0"
description = "A logger for fastapi"
authors = ["aldorax <climaxating@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
stackprinter = "^0.2.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = "^0.30.1"
sqlmodel = "^0.0.19"
redis = "^5.0.5"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
